// packages/plugin-assets-ui/src/common/types.ts
export interface IAssetCategory {
  _id: string;
  name: string;
  order: string;
  code: string;
  description?: string;
  attachment?: any;
  status: string;
  parentId?: string;
  createdAt: Date;
  assetCount: number;
  isRoot: boolean;
}

export interface IAsset {
  _id: string;
  name: string;
  type: string;
  categoryId: string;
  parentId: string;
  description: string;
  order: string;
  code: string;
  unitPrice: number;
  customFieldsData?: any;
  createdAt: Date;
  vendorId?: string;

  attachment?: any;
  attachmentMore?: any[];
  assetCount: number;
  minimiumCount: number;
  category: IAssetCategory;
  parent: IAsset;
  childAssetCount: number;
  vendor?: ICompany;
  kbArticleIds: string[];
  knowledgeData?: any;
}
// ...
export type IMovementType = {
  _id?: string;
  assetIds?: string[];
  description?: string;
  assets?: IMovementItem[];
  createdAt?: string;
  modifiedAt?: string;
  movedAt?: string;
  userId?: string;
  user?: any;
};

export type IMovementItem = {
  _id?: string;
  assetId: string;
  assetDetail?: IAsset;
  branchId?: string;
  departmentId?: string;
  teamMemberId?: string;
  customerId?: string;
  companyId?: string;
  movementId: string;
  branch?: any;
  department?: any;
  company?: any;
  customer?: any;
  teamMember?: any;
  createdAt?: string;
  sourceLocations: {
    branchId?: string;
    departmentId?: string;
    teamMemberId?: string;
    customerId?: string;
    companyId?: string;
    movementId: string;
    branch?: any;
    department?: any;
    company?: any;
    customer?: any;
    teamMember?: any;
  };
};